+++ src/otter_service_stdalone/grade_notebooks.py
@@ -1,7 +1,9 @@
 import asyncio
-import async_timeout
 from otter_service_stdalone import fs_logging as log, upload_handle as uh
 import os
+from otter.grade import main as grade
+from multiprocessing import Process
+from tornado.ioloop import PeriodicCallback
 
 log_debug = f'{os.environ.get("ENVIRONMENT")}-debug'
 log_count = f'{os.environ.get("ENVIRONMENT")}-count'
@@ -13,7 +15,6 @@ class GradeNotebooks():
         otter grader as well as a function for logging the number of 
         notebooks to be graded
     """
-
     def count_ipynb_files(self, directory, extension):
         """this count the files for logging purposes"""
         count = 0
@@ -22,7 +23,7 @@ class GradeNotebooks():
                 count += 1
         return count
 
-    async def grade(self, p, notebooks_path, results_id):
+    async def grade(self, p, notebooks_path, results_id, user_queue):
         """Calls otter grade asynchronously and writes the various log files
         and results of grading generating by otter-grader
 
@@ -44,61 +45,42 @@ class GradeNotebooks():
                            "",
                            "info",
                            f'{os.environ.get("ENVIRONMENT")}-count')
-            log.write_logs(results_id, "Step 5: Notebook Folder configured for grader",
+            log.write_logs(results_id, "Step 5: Notebook Folder configured and grading started",
                            f"Notebook Folder: {notebook_folder}",
                            "debug",
                            log_debug)
-            command = [
-                'otter', 'grade',
-                '-n', 'grader',
-                '-a', p,
-                notebook_folder,
-                "--ext", "ipynb",
-                "--containers", "10",
-                "--timeout", "1080",
-                "-o", notebook_folder,
-                "-v"
-            ]
-            log.write_logs(results_id, f"Step 6: Grading Start: {notebook_folder}",
-                           " ".join(command),
-                           "debug",
-                           log_debug)
-            process = await asyncio.create_subprocess_exec(
-                *command,
-                stdin=asyncio.subprocess.PIPE,
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE
-            )
+            p = Process(target=grade,
+                        kwargs = {
+                            "name": "grader",
+                            "autograder": p,
+                            "paths": (notebook_folder,),
+                            "containers": 10,
+                            "timeout": 300,
+                            "ext": "ipynb",
+                            "output_dir": notebook_folder,
+                            "result_queue": user_queue,
+                            "summaries": True,
+                        }
+                        )
+            p.start()
 
-            # this is waiting for communication back from the process
-            # some images are quite big and take some time to build the first
-            # time through - like 20 min for otter-grader
-            async with async_timeout.timeout(2000):
-                stdout, stderr = await process.communicate()
+            # Periodically check if the process is alive
+            def check_if_finished():
+                if not p.is_alive():
+                    log.write_logs(results_id, "Step 6: Grading: Finished",
+                                   f"{notebook_folder}",
+                                   "debug",
+                                   log_debug)
+                    log.write_logs(results_id, f"Grading: Finished: {notebook_folder}",
+                                   "",
+                                   "info",
+                                   log_error)
+                    periodic_callback.stop()
+                    return True
 
-                with open(f"{notebook_folder}/grading-output.txt", "w") as f:
-                    for line in stdout.decode().splitlines():
-                        f.write(line + "\n")
-                log.write_logs(results_id, "Step 7: Grading: Finished: Write: grading-output.txt",
-                               f"{notebook_folder}/grading-output.txt",
-                               "debug",
-                               log_debug)
-                with open(f"{notebook_folder}/grading-logs.txt", "w") as f:
-                    for line in stderr.decode().splitlines():
-                        f.write(line + "\n")
-                log.write_logs(results_id, "Step 8: Grading: Finished: Write grading-logs.txt",
-                               f"{notebook_folder}/grading-logs.txt",
-                               "debug",
-                               log_debug)
-                log.write_logs(results_id, f"Step 9: Grading: Finished: {notebook_folder}",
-                               " ".join(command),
-                               "debug",
-                               log_debug)
-                log.write_logs(results_id, f"Grading: Finished: {notebook_folder}",
-                               " ".join(command),
-                               "info",
-                               log_error)
-                return True
+            periodic_callback = PeriodicCallback(check_if_finished, 500)
+            periodic_callback.start()
+
         except asyncio.TimeoutError:
             raise Exception(f'Grading timed out for {notebook_folder}')
         except Exception as e: