+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.1.8
+
+#### Enhancements made
+
+* Remove Button for Submission Progress
+* Submission Progress in UI
+* Otter Grade: prints each notebooks summary
+
+## 1.1.2
+
+#### Enhancements made
+
+* Receiving updates while grading from Otter Grader
+* Otter Grade CSV to indicate which notebooks timeout
+
 ## 0.1.27
 
 #### Enhancements made